Address feebdback - split out RegisterLogin
