--- app/src/main/java/com/example/location_aware_personal_organizer/ComposeActivity.kt
@@ -1,83 +1,0 @@
-package com.example.location_aware_personal_organizer
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import com.example.location_aware_personal_organizer.services.Authorization
-import com.example.location_aware_personal_organizer.ui.theme.AppTheme
-
-class ComposeActivity : ComponentActivity() {
-    private fun onLogin() {
-        startActivity(Intent(this, MainActivity::class.java))
-    }
-    override fun onCreate(savedInstanceState: Bundle?) {
-        Authorization.getInstance();
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge();
-        setContent {
-            // A surface container using the 'background' color from the theme
-            AppTheme {
-                Surface(
-                    modifier = Modifier.fillMaxSize(),
-                    color = MaterialTheme.colorScheme.background
-                ) {
-                    RegisterLogin(onAuthentication={onLogin()})
-                    //MessageCard("Android", "test composable");
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun MessageCard(message: String, author: String) {
-    // Add padding around our message
-    Row(modifier = Modifier.padding(all = 8.dp)) {
-        Image(
-            painter = painterResource(R.drawable.ic_launcher_foreground),
-            contentDescription = "Contact profile picture",
-            modifier = Modifier
-                // Set image size to 40 dp
-                .size(40.dp)
-                // Clip image to be shaped as a circle
-                .clip(CircleShape)
-        )
-
-        // Add a horizontal space between the image and the column
-        Spacer(modifier = Modifier.width(8.dp))
-
-        Column {
-            Text(text = author)
-            // Add a vertical space between the author and message texts
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(text = message)
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    MessageCard("Android", "Test composable");
-}
